move new common bft classes under the bft package
